+++ Unity_ExperimentLibrary/Packages/com.maellacour.experiment-library/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.0] - 2025-02-25
+
+### Added
+
+- **Logs**: Add Simple log saver that save the player Log to participant folder.
+- **Logs**: Add Atom Log saver that save the player logs to participant folder base on atoms intReference.
+
+### Tests
+
+- **Logs**: Add tests for the log savers.
+
 ## [2.5.1] - 2024-12-02
 
 ### Fixed